Launcher Presenter removed (yeahhhh)
