+++ modules/app/src/main/scala/cards/nine/app/ui/launcher/jobs/uiactions/DockAppsUiActions.scala
@@ -1,10 +1,10 @@
 package cards.nine.app.ui.launcher.jobs.uiactions
 
 import android.support.v4.app.{Fragment, FragmentManager}
+import cards.nine.app.ui.commons.UiContext
 import cards.nine.app.ui.commons.ops.UiOps._
-import cards.nine.app.ui.commons.{AppUtils, UiContext}
 import cards.nine.app.ui.components.layouts.tweaks.DockAppsPanelLayoutTweaks._
-import cards.nine.commons.services.TaskService
+import cards.nine.app.ui.launcher.LauncherActivity._
 import cards.nine.commons.services.TaskService.TaskService
 import cards.nine.models.DockAppData
 import cards.nine.process.theme.models.NineCardsTheme
@@ -16,16 +16,7 @@ class DockAppsUiActions(val dom: LauncherDOM)
     fragmentManagerContext: FragmentManagerContext[Fragment, FragmentManager],
     uiContext: UiContext[_]) {
 
-  case class State(theme: NineCardsTheme = AppUtils.getDefaultTheme)
-
-  private[this] var actionsState = State()
-
-  implicit def theme: NineCardsTheme = actionsState.theme
-
-  def initialize(nineCardsTheme: NineCardsTheme): TaskService[Unit] =
-    TaskService.right {
-      actionsState = actionsState.copy(theme = nineCardsTheme)
-    }
+  implicit def theme: NineCardsTheme = statuses.theme
 
   def loadDockApps(apps: Seq[DockAppData]): TaskService[Unit] =
     (dom.dockAppsPanel <~ daplInit(apps)).toService
